chore: update readmes
These READMEs are included in the auto-generated docs sites. The typedoc
tool we use for TypeScript looks for the nearest README.md file, so it
ended up using the one at the root of each module (i.e.
`cdp-agentkit-core/README.md`), which included instructions for both
languages.

This commit removes the general README (i.e.
`cdp-agentkit-core/README.md`) in favor of language-specific ones, which
are now included in the docs sites.